src/test/java/ch/devcon5/sonar/plugins/mutationanalysis/MutationAnalysisPluginTest.java
/*
 * Mutation Analysis Plugin
 * Copyright (C) 2015-2018 DevCon5 GmbH, Switzerland
 * info@devcon5.ch
 *
 * This program is free software; you can redistribute it and/or
 * modify it under the terms of the GNU Lesser General Public
 * License as published by the Free Software Foundation; either
 * version 3 of the License, or (at your option) any later version.
 *
 * This program is distributed in the hope that it will be useful,
 * but WITHOUT ANY WARRANTY; without even the implied warranty of
 * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
 * Lesser General Public License for more details.
 *
 * You should have received a copy of the GNU Lesser General Public License
 * along with this program; if not, write to the Free Software Foundation,
 * Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
 */

package ch.devcon5.sonar.plugins.mutationanalysis;

import static org.junit.Assert.assertFalse;

import org.junit.Test;
import org.junit.runner.RunWith;
import org.mockito.InjectMocks;
import org.mockito.Mock;
import org.mockito.runners.MockitoJUnitRunner;
import org.sonar.api.Plugin;
import org.sonar.api.SonarRuntime;

@RunWith(MockitoJUnitRunner.class)
public class MutationAnalysisPluginTest {

    @InjectMocks
    private MutationAnalysisPlugin subject;

    @Mock
    private SonarRuntime sonarRuntime;

    @Test
    public void testDefine() throws Exception {

        Plugin.Context context = new Plugin.Context(sonarRuntime);

        subject.define(context);

        assertFalse(context.getExtensions().isEmpty());

    }

}
